feat: migrate PMI/GD&T placement to the IR — the last feature epic (#200/#208)
The PMI annotation pass moves onto the one path: `PmiFeature` (IR) + `render_pmi`
(renderer); the engine's `_annotate_pmi` and `annotations/pmi.py` are deleted. This
closes the ADR-0008 convergence — the orchestrator is now build → plan → render +
the shared section/table passes, with no per-feature engine pass left.

PMI is *pre-authored* (label/value/tolerances/geometry are set by the CAD author in
the STEP AP242 file), so unlike the geometric features it does not flow through the
dimension planner — there is no convention or view to derive. `PmiFeature.parameters()`
is therefore empty and `render_pmi` consumes the features directly from the model (the
same way `render_slots` reads `SlotFeature`s). This keeps PMI on the one path without
pretending it benefits from the planner.

- ir.py: `PmiFeature` (kind/value/label/dominant_axis/ref_bbox/ref_pts).
- detect.py: `build_part_model(..., pmi=)` re-homes the extracted `PmiRecord`s as
  `PmiFeature`s. (`extract_pmi` in pmi.py — the STEP-side extraction — is unchanged.)
- from_model.py: `render_pmi` — a faithful port of `_annotate_pmi`'s placement
  (witness-from-bbox, bore leaders, strip allocation), sourced from the IR.
- orchestrator: calls `render_pmi(dwg, _model, a)` under the same `pmi_mode == "annotate"`
  gate; `_annotate_pmi` import + `annotations/pmi.py` removed.

Verified: CTC-01 AP242 (`pmi="annotate"`) → 8 `pmi_` annotations, lint clean, visual
parity with the engine pass; report/off modes add no annotations (planner/lint never
see PmiFeatures — `parameters()` is empty). test_pmi 15 passed; full fast suite 585
passed / 1 skipped; ruff(0.15.17)/format/mypy clean.

Closes #200, #208.
