+++ src/draftwright/model/__init__.py
@@ -32,6 +32,7 @@ from draftwright.model.ir import (
     HoleFeature,
     PartModel,
     PatternFeature,
+    PmiFeature,
     RotationalFeature,
     SlotFeature,
     StepFeature,
@@ -57,6 +58,7 @@ __all__ = [
     "HoleFeature",
     "PartModel",
     "PatternFeature",
+    "PmiFeature",
     "RotationalFeature",
     "PlannedDimension",
     "SlotFeature",
